유저네임 표시 수정
마이페이지를 열면 유저네임이 표시되도록 수정했습니다.
